+++ R/tidytargets.R
@@ -10,6 +10,7 @@
 #' @importFrom magrittr %>%
 #' @importFrom dplyr tibble group_by summarise
 #' @importFrom rlang parse_expr
+#' @importFrom qs2 qs_save qs_read
 #' @importFrom targets tar_target_raw tar_resources tar_resources_crew tar_option_get tar_make tar_meta tar_config_get tar_option_set tar_cue
 #' @importFrom tarchetypes tar_quarto_raw
 #' @importFrom callr r